Reorganize Project
This reorganizes the project to place the caching plugin as the index.js
and then also places cache and key for file in a lib directory.
