+++ lib/key-for-file.js
