+++ test/test.js
@@ -12,7 +12,7 @@ var fs = require('fs');
 var mkdirp = require('mkdirp');
 var path = require('path');
 var Builder = require('broccoli').Builder;
-var Filter = require('../filter.js');
+var Filter = require('../index.js');
 var minimatch = require('minimatch');
 var rimraf = require('rimraf').sync;
 var walkSync = require('walk-sync');
